refactor: organize functions and types related to format search params
