+++ apps/storefront/src/types/prisma-order-by.ts
@@ -1,0 +1,3 @@
+import prisma from "@/lib/prisma";
+
+export type PrismaOrderBy = Parameters<typeof prisma.product.findMany>['0']['orderBy']
