+++ apps/storefront/src/types/search-params.ts
@@ -1,0 +1,1 @@
+export type SearchParams = Record<string, string | string[]>
